For #16615 & #11191: re-enabled ReaderViewTest and added a UI smoke test
